Fixed a few bugs, found some others, improved the documentation.
More functions have been documented in "liblisp.md", mainly maths ones.

A cause for a segmentation fault was found and fixed; the jump buffer was
not restored correctly, this might need some more investigation. The
fault occurred when using the line editor and inputing invalid data or
throwing errors from within the lisp environment.

A memory leak was found, but not fixed. This can be triggered when using
the line editor and inputting an invalid line, such as

        (

With no terminating bracket.