More comments, changed "type?" to "type-of", fixed a few "XXX"
The "type?" subroutine, or primitive, was removed and replaced with "type-of"
which is more flexible and smaller.

A few new comments where added, some moved around, and more documentation was
added. There is still a lot more documentation to do however.

The project is nearing completion (and its self imposed line limit of 3000
lines for the core interpreter).

A few of the 'XXX' comments were removed, these represent a known bug or
missing functionality. There were three removed;

1) Coercion from list to string. This is not going to be implemented any
   time soon, the functionality can be implemented in pure lisp as well.

2) "subr_read" did not catch its longjmp that it could throw. It does not.
   The longjmp handling code is very ugly however.