Added "subr_eval" catch, fixed subr_seek
Added a catch for for subr_eval, as in subr_read, this swaps out the longjmp
buffer used for a temporary one in case eval needs to RECOVER or HALT. The
current semantics with HALT do not quite work correctly however, but it works
better than it previously does. subr_eval can still be used to recurse
infinitely (well until a stack overflow happens) which is an error, it could be
made into a special primitive like "if" or "cond" instead.

subr_seek had a fix that was incorrect, "whence" should have been checked
instead of the offset. The valid values for whence are SEEK_SET, SEEK_CUR and
SEEK_END, represented by *seek-set*, *seek-cur* and *seek-end* in the actual
interpreter.