Adding complex arithmetic functionality as a test
This functionality that deals with complex numbers is for test purposes only,
it is not complete nor will it be tested that much, only to see how feasible it
is.

In this commit, which gives very basic complex functionality, I have:

* Support for c99 complex math functions
* Removed documentation that refers to normal, non-complex math functions
* Added support for printing complex numbers
* Tested that (monte-carlo-pi 400) works with the changes
* Fixed mk() but not tested it under all cases
* Some very, very basic tests.

Still to do:

* Input of complex numbers instead of making them by doing (sqrt -1).
* If the imaginary part is zero it should not print it out.
* Some very minor testing.
* Add documentation for complex functions (very low priority).
* Figure a way to get this working under Windows (very low priority).

It has been easier to add than I would have thought.