Documentation, minor bugs, more notes, balance, io
The documentation continues to be improved, in doing so some minor problems
with the interpreter were discovered and noted.

The IO String output functions io_puts and io_putc should now grow their
buffers when the file end has been reached like a file would be. This has not
been test yet.

The PRNG now discards the first N numbers.

A name collision in "init.lsp" was removed.

The value entries in a hash table that are printed out should be printed out as
escaped string, but they are not, this has been noted but not fixed.

subr_hinsert can now handle both strings and symbols as the key value.