Removed files, better error msg in main, doc, other
Removed unneeded files in the "exp/" folder, they are linked else where in the
documentation.

Improved the error messages in main.c for functions that are added externally,
using a recently improved API.

The documentation continues to be expanded.

More comments were added, some removed.

Coercion from strings to symbols was improved.