Makefile, assoc w/ hashes, close, comments, docs
Some of the larger comments have been moved to the "liblisp.md" document.

Many more smaller comments have been added documenting the internal data
structures used.

The assoc function now works with hashes as well, although the internals have
yet to capitalize on that.

The "subr_close" function has been implemented with another bitfield in the
"cell" object that indicates that an objects data field has been finalized even
if the object has not be collected itself.

The makefile should automatically checkout libline if it has not been, this
hash not been tested and it surely can be improved.

A note was added in the "make.bat" file for building and running under Windows,
that file should definitely be improved upon to be more like the proper make
file, as much as is *possible*.