Fixed one minor leak, removed debugging option.
Fixed one minor memory leak which was caused by a clobbered pointer, which was
partially caused by a clobbered pointer and partly a misplaced free. The other
one has not been fixed, but it has been pinned down, another pointer is
clobbered by longjmp, I still need a way to resolve this though.

I have removed the debugging option from the lisp_main function, it was not
very useful. Although now there is no way to tell when a new file is being
loaded.