docs, signal handling, HOME handling, args, API
There has been minor addition to the documentation in the forms of new
examples.

If the HOME environment variable is set it will attempt to write the history to
"HOME/.list", if it cannot it will write the history to ".list" in the current
working directory. I should check whether the history is actually being
written...

Signal handling has been moved to outside of the interpreter, a function is
provided that a signal handler or another thread can use to tell a lisp
environment that something has happened and it should stop evaluating. It only
works with the line editor at the moment, so if it is disabled then SIGINT will
only be caught once and not reset. This required changing the API a little to
add a function called "lisp_set_signal".

New options were added to the example REPLs options; "--" and "-i" which act in
the same way, it allows you to process files with odd names that look like
command line flags and options themselves.

More comments have been added to the code as well.

A memory leak in libline has been identified, it is minor and it will leak a
constant amount, but it should be fixed.