arg bug, libline mem leak, make.bat, hist check
A bug recently introduced with a new command line option was fixed, when "--"
or "-i" was passed in it did not check whether the next argument in the list
existed.

A memory leak in libline was fixed which was due to a function to free the
history not being called atexit().

"make.bat" has been improved so it acts a bit more like a "makefile", however
it would be better to make the normal "makefile" also work on Windows. This
would be simplified if libline was ported to Windows as there would be less to
configure.

A check to see if the history file can be saved to has been added.