Removed unneeded lisp code, limit print depth, doc
Some lisp code relating to "slurping" files up for reading was removed as
"get-delim" with EOF as the delimiter can replace it largely.

The printing code has a recursion depth limit placed on it and can now print
out user defined function which have a printing function associated with them
(although this has not been tested).

Some more comments were added as well.