doxygen, comments, new options, getdelim, hash, match, API
The simple match function was improved on a little, it still needs overflow
checks however to prevent a stack overflow. It has different syntax and now
supports escape characters for the matching characters.

io_getdelim was improved so it grows by a factor of 2 each time instead of by a
constant amount.

There are some more comments about potential branches, such as a cut down
version.

The doxygen.conf file has been removed, the default documentation generated by
doxygen works fine for this project, anything more is not really needed. I do
need to fix many warnings generated by doxygen, some are valid, others are not.

A minor inconsistency in the API was fixed, lisp_add_subr and lisp_add_cell
have similar function signatures.

subr_hdefined has been added which allows a user to check whether a value has
been defined in a hash table or not.

A new command line option has been added, this option "-H", halts the
interpreter when an error is encountered, instead of continuing to process
options. This should probably be the default behavior unless we are in line
editing mode.

New ideas:
        * Make a branch for a cut down version with no floats and hashes,
          just get rid of as much as possible.
        * Test out different data structures in the interpreter. I really
          want to reduce the amount of allocations performed, so this means
          linear probing for hashes, dynamic arrays the garbage collection
          list of all allocations and using the "struct hack" for string
          and symbol objects.

A help section has also been added to the makefile, this makefile should be
commented more.