subr_hdefined removed => hash semantics changed
subr_hdefined was removed, along with the documentation for it, and the
documentation and code updated so that subr_hdefined is not needed. It behaves
more like assoc, it returns a cons of the key and the value if the key was
found in a hash and nil when nothing has been found.