flambda mapcar, documentation, ';' comments
A flambda version of mapcar has been added, it does not quite work as intended
as it cannot be used within functions without caveats.

The documentation has been improved.

';' can now be used as a comment as well

A bug has been fixed with "all-symbols" subroutine printing, which was caused
by complications of the hash implementation.