Added USE_TCC, user-def API changes, rename lisp funcs
A libtcc interface has been added, but it is not a mature feature, it should
not be used. This feature would allow the lisp interpreter to define

As a consequence of adding libtcc some changes to the API have been added for
user defined functions to make up for some deficiencies with it.

A few lisp functions have been renamed.

The documentation has been updated to reflect these changes.