gensym, asserts, tr, moved ctype funcs, match, icbins
New lisp code has been added for gensym which can be used to generate
globally unique symbols.

Lots of error checking code has been replaced with asserts, which naturally can
be removed from the build. These asserts should only be triggered due to
programmer error, most likely because some out of memory condition was not
checked for handled correctly (the interpreter should signal the calling
functions that there is a problem so that it can exit gracefully).

The "tr" subroutine has been added, but it needs some work, it is useful for
processing text, a "diff" routine should be added next, then eventually a "sed"
engine.

The ctype.h functions have been moved from main.c to subr.c as has been
planned.

An interesting scheme interpreter and compiler has been added called "icbins",
this is quite an old version of the interpreter. A newer version by the same
author can be found at <https://github.com/darius/ichbins>, it is very elegant.

The match function now contains an ASSERT (not assert) which is a very crude
way of prevent a stack overflow, this should be fixed, but will do for now.

The interpreter is nearing completion.