Fixed bug that caused segfault on handling of dotted pairs
Until this fix, the following would cause a segfault as evlis did not handle
the case of there being a dotted pair passed in as arguments.

        (+ 1 . 2)

Also of note is that the length command does not work correctly with dotted
pairs and lists ending in dotted pairs, although it is not clear what the
behavior should be.