bitfield, modules, renames
Added bit field manipulation code, this is not used in the interpreter, but is
useful, it might be added to the interpreter functionality later.

I have renamed various functions to be more consistent with the APIs naming of
other functions, this work should be continued.

An example module for interfacing with sqlite3 has been added, but not
completed.