Move line editing and math functions to modules
The line editing and math functions have been moved into various modules, the
module system means the main interpreter will not have to become bloated in
order to add new functionality; it can be tried and loaded in modules instead.

There are still problems with the interaction between the module system and
garbage collection.