Fixed dlclose on exit, subr_define_eval, init.lsp
main.c now creates a linked list of dynamically loaded modules to close on
exit, there are still memory leaks due to one of the modules (SQLite3 module)
that need resolving and with the Tiny C Compiler module if it is used.

subr_define_eval has been added, this allows computed symbol names to be
defined, this subroutine could probably be used in conjunction with
F-Expressions to replace the special form "define", but this has not been tried
yet.

The new subr_define_eval has been used in the init.lsp code for module loading.