better modules, validation, makefile, x11 module, comments, ...
The modules have undergone various minor improvements.

A new file has been added that adds a generic way to validate built in
subroutines, it is hoped that this along with a doc-strings will provide a
minimal set of online documentation for the interpreter. In the mean time the
validation routine, whilst primitive at the moment, is being used to reduce the
amount code in the "subr.c" file. The check string should be added to the
X-Macro of all the subroutines along with the length check, this would allow
the validation to be done in a single place in the "eval.c" file.

The validation function works be accepting a format string that describes the
length of the argument list and the types of each argument that are valid to
the subroutine.

This method also allows for the validation checks to be disabled, if
performance is a concern, the previous and more ad hoc way did not allow for
this.

Not all functions have been converted yet as the "lisp_validate" function still
needs work.

An x11 window module has been added, but not expanded on, it just opens a
window at the moment when run with no functionality for event handling or
drawing. This will be added soon, but the module is not intended to provide a
vast amount of functionality.

The makefile has been improved, and could do with more improvement still to
allow it to be used in a Windows environment, for the moment it is strictly
tied to the Linux platform with the "make.bat" file for Windows (which is
usually broken in some minor way). The modules do not have to be installed to
be run now, instead the LD_LIBRARY_PATH variable is set to the "module/"
directory.

Many more @todo comments have been added, as well as comments in other files.