bugs, renames, removed functionality, validation
A minor type checking bug was fixed in coerce.

The functionality of the interpreter has been changed:
* letrec was renamed to let.
* let* was removed.
* trace levels (but not tracing) and control over the garbage collector
  (apart from forcing a collection) have been removed.
* The error special form has been removed
* Now incorrect sections of liblisp.md have been removed and/or changed.

The lisp_validate routine has been used in more places where possible.

This is one of the only recent commits to both remove functionality that was
deemed not needed and reduce the overall line count.