Function renaming, file removal, looping constructs, ...
Many functions have been renamed so as to be consistent with the
semi-established naming convention in the library, for example going from
"mkint" to "mk_int" which accounts for most of the files having been changed.

Some unneeded code has been removed, as have unneeded files.

Looping constructs have been added, although the semantics may need to be
worked out.

A bug has been identified, the length of a list at its head it calculated
correctly, but it is not correct further through the list, this is not usually
visible to the user.

Some functionality has been removed and subroutines have had functionality
removed.

The validate function has been added as a subroutine.