asserts, preparation for error handling / docstrings, validate
More asserts have been added to the program, I would like to add many more to
as they form a way of aiding in the assurance that the program is correct and
they form a way of documenting the code that cannot be incorrect or the program
will fail. Ideally on failure a stack trace would be printed out, but this
cannot be done portably. They can be removed if they pose too much of a
performance penalty, naturally.

Preparation for adding a generic way of handling errors had been added, as well
as documentation strings, for built in subroutines. The table will have to be
completed, as well as modifications made to the rest of the interpreter to
accommodate it.

The "validate.c" file has been changed to use an X-Macro table to generate code
to ensure that two function stay in sync and to reduce repetition.