const, better validation, better match, ...
The simple "match" C routine has been improved, so it does not exit the
interpreter but uses longjmp for error recovery.

"const" has been put on functions that do not modify strings passed to them.

Input validation of built-in lisp subroutines has been moved so it is checked
in "eval.c" and is based on a validation string input in an X-Macro, the error
reporting offered by this is very much lacking and needs to be improved.

A few new functions have been added to accommodate these features. These new
features need to be extended to include doc-strings and use more throughly
throughout the interpreter.