Introspection, doc-strings, input/output/error ports, ...
Some primitives have been added for introspection, I would like to add more
when I can think of them, the ones that have been added are for querying
functions (both user defined and built in functions).

The input, output and error ports that the interpreter is currently using have
been made accessible to the interpreter; the interpreter reads in code which is
handled in with a port, once it has finished reading in a file it will close
this port and potentially open up a new one, the input port available from
within the interpreter gets updated so that it refers to the correct file.

"doc-strings" now work for primitive subroutines, but not for user defined
functions (those defined with lambda or flambda), although the error reporting
mechanism currently does not make use of this.