renaming, monte.c, makefile, man pages, ...
Many of the C API functions have been renamed, mainly to do with accessing
variables. As an example:

        ioval  -> get_io
        strval -> get_str

This is in keeping with how the other functions are named such as

        mk_io
        is_io

        mk_str
        is_str

mk_subr_long has been removed and the functionality put into mk_subr. There are
other types and functions that have likewise been changed.

The man pages have been updated to be more in the style of other man pages and
the makefile updated to install "liblisp.3" correctly.

"monte.c" has been added, it is a small C program that approximates the value of
Pi via Monte-Carlo methods, there is a function in "test.lsp" that does the
same thing. This provides a method to benchmark the interpreter, roughly, at
the moment it does *very* poorly and needs improvement. All changes aimed at
improving speed can be checked against this program (as well as with the usual
methods).

As usual there have been other minor changes throughout as well.