Rename, module system, init.lsp, makefiles, new subrs
The makefile now compiles more cleanly on Windows (using mingw32) although
there is still work to be done with in, not all of the makefiles features are
available on Windows at the moment (such as install/uninstall/dist). The
makefile still works perfectly under Linux (on Debian 7/8).

init.lsp has been gutted at moved into the base.lsp, it now contains only the
minimum needed to evaluate files.

Some functions have been renamed to be more consistent with the naming
convention, this still needs to be done with all of the modules.

A few new subroutines have been added.

I need to think about the build system and the module system some more,
currently loading the compiled modules takes too long, and it will only get
worse as more modules are added.

Also, the number of @todo and @bugs are starting to pile up.