Sped up interpreter, improved error reporting, ...
The interpreter has been sped up quite a lot by using a hash for the top level
for defined values. It is roughly be a factor of ten.

The error reporting has been improved with colorized messages.

"-DFORTIFY_SOURCE=2" and "-fstack-protector" has been added to the makefile.

Garbage collection is turned off during initialization.

There are other minor changes as well.