Added compilation skeleton, environment, ...
A skeleton for "compile", this is will not be a true compiler and will be
removed if it does not improve speed too much, it will partially evaluated
expressions so symbol lookup does not have to be done and expressions are
rewritten to be more efficient.

"environment" is not longer a special symbol / special form, the current
environment is tucked away for future use.

There have been other minor changes throughout the interpreter as well.