Improved makefile, unit tests framework, added mods.lsp, ...
The makefiles have been improved again and changes make to the interpreter to
make it more compatible with Windows, the library still compiles perfectly
under Linux.

Two new directories have been added to the rpath of the "lisp" executable that
is built, this has not been tested under Windows. It adds the current and
"mod/" directories to the shared library search path.

A unit test framework has been added, it has uncovered some assumptions I have
made in the code already. The "util.c" nearly has full coverage, it might be
quite difficult to test the interpreter itself sensibly however.

I have found a bug in "read.c" and how it interacts with the rest of the
interpreter when it fails to read in a string correctly, this will be
documented and then fixed, it produces quite confusing behavior.

As usual, there are other minor changes throughout.

Also of note, files in the "mod/lisp" directory are not been added with
the "git add ." command, this should not happen, although the entry "lisp" is
in the ".gitignore" file, (see <https://git-scm.com/docs/gitignore>).