foldl, makefile, init.lsp, monte.pl, fixes, bug...
The makefile has been made to be more portable, and more functionality is
available on Windows, although the Windows built is still very likely to be the
first to break as it gets tested less often. Install/Uninstall now work. The
makefile is starting to get unwieldy, a better way of organizing it should be
found.

The primitive "foldl" has been added, the first of hopefully many "higher"
order functions.

init.lsp has been made to be more portable also.

"monte.pl" has been, to add a bench mark of another interpreter.

A bug has been fixed in the ">" and "<" routines, (< "a" 1) would fail.

There have been minor fixes throughout as usual. Mostly related to warnings on
Windows.

A bug has been found in the evaluation of dotted pairs (again):
        (eval '(quote . 2))
Causes an assertion to trigger, this should be fixed.