docstrings, unit SIGABRT catch, comments, notes, stats, ...
The docstring format has been standardized, although not implemented for all
submodules functions. The format is:
         <name> <seperator> <file> <seperator> <description>
The idea being that this can be split and printed out.

The unit test suite can now catch SIGABRT (and potentially other fatal signals)
raised by what it is testing, the test suite logs this as a failure and
continues onto the next test. One of the failing tests has been added that
causes a SIGABRT. The test suite is almost ready to be split off and turned
into a library in itself, perhaps in "util.c".

More comments and "@todo"s have been added throughout as well as notes.

Some statistics functions have been added to the interpreter.