slre, docstrings, (eval (cons quote 0)), mk_list, makefile, ...
The makefiles have been improved, for Windows, functionally they
nearly reach parity with the Unix side.

The internal docstring representation has been improved, in preparation for
lambdas and f-expressions to allow docstrings.

A few bugs have been caught and fixed (by the use of clang).
And "(eval (cons quote 0))" now no longer causes an assert() to be thrown,
instead correctly reporting an error.

If the first argument of an expression is an expression, it is evaluated and
returned. This now works:
        ((quote if) nil 'a 'b) => 'b

mk_list has been added, this allows the creation of lists without stacking
calls to cons() in the C code.

"data.lsp" has been added, for sets of data and the functions that operate
on them.

The project <https://github.com/cesanta/slre> could be used as a replacement
for the internal regex implementation, it does quite a bit, previous versions
were under a more liberal license, see:
        <http://slre.sourceforge.net/>
        <https://code.google.com/p/slre/>

Some code has been moved around (accounting for a large diff in this commit).