Improved hash library, colorized errors, basic completions, gc bug
The hash foreach function has been improved greatly, allowing resumption and a
full foreach to be done on the hash.

The colorization of errors in "eval.c" has been improved.

A gc bug has been fixed, sort of, a NULL can be added to the free list when an
error occurs in the input which triggered an assert. In the case of a NULL in
the free list the function simply returns.

Very basic completions have been added. It needs a lot of work doing to do to
make it aware of the lisp syntax.

The libline library has had its completion callback improved to give the
current cursor position in the line so more advanced completions can eventually
be added.