Improved makefiles, reorganized project, ...
The makefile for the project has been improved, the makefile in the mod/ folder
is now included in the main make file, there could still be a lot of
improvements. The changes have not been tested on Windows, only on Linux. The
changes avoid the use of recursive make and improve the makefile generally
speaking.

The project has been reorganized, the modules liblisp_diff and liblisp_tsort
have been merged in anticipation of adding more functionality to them.

More test cases have been added to "unit.c".

More constants have been added to the interpreter, for floats and integers.

There have been other minor changes as well.