merged makefile and config.mk, documentation (minor commit)
There is very little in this commit. Makefile and config.mk are being merged.
Perhaps a script, a perl script maybe, could be used to automatically generate
a better config.mk. For the moment these files will stay merged.

The documentation in the header file has been improved in minor ways.