SQL, XML, parse/print hashes, memory leak info, comments
The SQL and XML modules have been improved, the XML module is now
semi-functional and not tested that much. XML can now be parsed in and written
out, but there is still a lot to do and decisions as to how the XML should be
parsed and printed.

Hashes can now be read in and printed out, they are now part of the syntax,
this parsing can still be turned off if desired. Hashes can still be created
using in the usual fashion with hash-create.

The memory leak when loading modules appears to be spurious, possibly and
caused by an interaction with dlopen and valgrind, and not with the SQL module.

More function in "liblisp.h" have been documented.

More @todos have been added, I should clear them up soon.