More renames, modules interface, curl & PCRE, misc, trace
More enumerations and structures have been renamed, such as FIN becoming
IO_FIN.

The module interface has been improved, a convenience function has been added
for adding all the subroutines in a dll.

The curl module has had some basic functionality added to it, although nothing
to do with the main functionality curl provides, which will have to be added
later.

The PCRE library has been improved, the interface is pretty much how I want it,
although there are bound to be a lot of bugs and edge cases in it.

Moved *version*, *commit* and *repository-origin* to the library from "main.c",
they will be set to "unknown" if the build system does not supply defines for
these.

The previous trace mechanism has been removed in favor of controlling log
levels from within the interpreter (as well as by command line switches, which
a running interpreter can override).