Testing new hash function, lisp copy, makefile
This new hash function is a lot more customizable, hashes can be created with
custom callback functions for freeing, comparing and hashing values, making the
hash library much more useful and generic.

A function for copying lisp data structures, where possible, has been added,
and another function for copying lisp hashes.

The makefile has been updated for portability reasons.

There appears to be a serious bug somewhere in the project, I need to find out
if it is caused by my new code, hence committing this to a branch first.

The symptom is that "even?" appears not to compile, for some reason, when a few
more symbols are defined. Perhaps it is getting lost in the hash? Or not
checked for? Somehow?