Adding destructive operator test code, bug still present
A destructive + and - have been added, they modify there first argument instead
of creating a new lisp integer or float cell, it speeds things up a little, the
changes would need to be propagated further in order to fully test the speed
up.

The bug mentioned in the previous commit is still present.