Reformat code, car/cdr, function names, prolog, makefile, destructive +/- removed
A lot of the lisp code has been reformatted so it is easier to read, and a lot
of function names have been changed to be more consistent, and so they look
nicer when using the "a.b" syntax. Most of the function names that have been
renamed are of the predicate form, being renamed from "nil?" to "is-nil" for
example, this looks better when doing "is-nil.x", versus "nil?.x" (which looks
weird).

car/cdr have been made to be more forgiving in terms of what they receive to
make error checking easier; (car nil) and (cdr nil) both return nil.

A buggy prolog interpreter has been added that has been half ported from the
original (written in XLISP).

The makefile has been changes, mostly style changes, but a few things have been
refactored to be easier to maintain in the future.

The destructive +/- have been removed.

An example of a closure has been added, although it has to be achieved manual
with a "copy".