Makefile now should work on Windows and Linux
The makefile has (with difficulty) been made sort of portable between Windows
and Linux.

Of note, recursive data structures are still not quite handled correctly, these
data structures can be created with "set-car" and "set-cdr", a patch was
introduced in previous commits to print out a reference instead - this did not
create the problem, only highlights certain cases, but it is not completed.